Removed unused code related to Connect AWS account
