+++ LocationServices/LocationServices/Scenes/Explore/ViewModel/ExploreViewModel.swift
@@ -17,9 +17,9 @@ final class ExploreViewModel: ExploreViewModelProtocol {
     
     var delegate: ExploreViewModelOutputDelegate?
     
-    var awsLoginService: AWSLoginService! {
+    var awsAuthService: AWSAuthService! {
         didSet {
-            awsLoginService.delegate = self
+            
         }
     }
     let locationService: LocationServiceable
@@ -84,16 +84,6 @@ final class ExploreViewModel: ExploreViewModelProtocol {
         }
     }
     
-    func login() {
-        Task {
-            try await awsLoginService.login()
-        }
-    }
-    
-    func logout() {
-        awsLoginService.logout()
-    }
-    
     func loadPlace(for coordinates: CLLocationCoordinate2D, userLocation: CLLocationCoordinate2D?) async {
         var biasLocation = userLocation
         if let mapCenter = UserDefaultsHelper.getObject(value: LocationCoordinate2D.self, key: .mapCenter) {
@@ -119,19 +109,5 @@ final class ExploreViewModel: ExploreViewModelProtocol {
         let welcomeShownVersion = UserDefaultsHelper.get(for: String.self, key: .termsAndConditionsAgreedVersion)
         let currentVersion = UIApplication.appVersion()
         return welcomeShownVersion != currentVersion
-    }
-}
-
-extension ExploreViewModel: AWSLoginServiceOutputProtocol {
-    func loginResult(_ user: AWSUserModel?, error: Error?) {
-        guard let user = user else { return }
-        let presentation = ExplorePresentation(model: user)
-        UserDefaultsHelper.save(value: presentation.userInitial, key: .userInitial)
-        delegate?.loginCompleted(presentation)
-    }
-    
-    func logoutResult(_ error: Error?) {
-        delegate?.logoutCompleted()
-        print("Logout Successfully")
     }
 }
