--- LocationServices/LocationServices/Scenes/Login/Presentation/LoginPresentation.swift
@@ -1,8 +1,0 @@
-//
-//  LoginPresentation.swift
-//  LocationServices
-//
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: MIT-0
-
-import Foundation
