+++ LocationServices/LocationServices/Scenes/Settings/Controller/SettingsVC.swift
@@ -32,14 +32,6 @@ final class SettingsVC: UIViewController {
         return tableView
     }()
     
-    private lazy var settingsButton: SettingsButtonView = {
-        let view = SettingsButtonView()
-        view.isUserInteractionEnabled = true
-        let tap = UITapGestureRecognizer(target: self, action: #selector(buttonAction))
-        view.addGestureRecognizer(tap)
-        return view
-    }()
-    
     var viewModel: SettingsViewModelProtocol! {
         didSet {
             viewModel.delegate = self
@@ -59,40 +51,21 @@ final class SettingsVC: UIViewController {
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         viewModel.loadData()
-        updateLogoutButtonVisibility()
     }
     
     private func setupNavigationItems() {
         navigationController?.isNavigationBarHidden = !UIDevice.current.isPad
         navigationItem.backButtonTitle = ""
     }
-    
-    @objc func buttonAction() {
-        switch UserDefaultsHelper.getAppState() {
-        case .loggedIn:
-            self.viewModel.logOut()
-        case .customAWSConnected:
-            self.viewModel.disconnectAWS()
-        default:
-            print("no action required")
-        }
-    }
    
     private func setupViews() {
         self.view.addSubview(headerTitle)
         self.view.addSubview(tableView)
-        self.view.addSubview(settingsButton)
     
         headerTitle.snp.makeConstraints {
             $0.top.equalTo(self.view.safeAreaLayoutGuide)
             $0.leading.equalToSuperview().offset(Constants.horizontalOffset)
             $0.trailing.equalToSuperview()
-        }
-        
-        settingsButton.snp.makeConstraints {
-            $0.height.equalTo(72)
-            $0.leading.trailing.equalToSuperview()
-            $0.bottom.equalTo(view.safeAreaInsets).offset(-16)
         }
         
         tableView.snp.makeConstraints {
@@ -102,7 +75,7 @@ final class SettingsVC: UIViewController {
             } else {
                 $0.leading.trailing.equalToSuperview().inset(Constants.horizontalOffset)
             }
-            $0.bottom.equalTo(settingsButton.snp.top)
+            $0.bottom.equalToSuperview()
         }
     }
     
@@ -111,23 +84,6 @@ final class SettingsVC: UIViewController {
     }
     
     @objc private func authorizationStatusChanged(_ notification: Notification) {
-        self.updateLogoutButtonVisibility()
-    }
-    
-    private func updateLogoutButtonVisibility() {
-        // show logout & disconnect button
-        DispatchQueue.main.async {
-            switch UserDefaultsHelper.getAppState() {
-            case .loggedIn:
-                self.settingsButton.setButtonState(settingsButtonState: .logout)
-                self.settingsButton.isHidden = false
-            case .customAWSConnected:
-                self.settingsButton.setButtonState(settingsButtonState: .disconnect)
-                self.settingsButton.isHidden = false
-            default:
-                self.settingsButton.isHidden = true
-            }
-        }
     }
 }
 
@@ -136,9 +92,5 @@ extension SettingsVC: SettingsViewModelOutputDelegate {
         DispatchQueue.main.async {
             self.tableView.reloadData()
         }
-    }
-    
-    func logoutCompleted() {
-        updateLogoutButtonVisibility()
     }
 }
