--- LocationServices/LocationServices/Scenes/Tracking/Subviews/Tracking History/View/TrackingHistorySectionHeaderView.swift
@@ -1,71 +1,0 @@
-//
-//  TrackingHistorySectionHeaderView.swift
-//  LocationServices
-//
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: MIT-0
-
-import UIKit
-import SnapKit
-
-final class TrackingHistorySectionHeaderView: UITableViewHeaderFooterView {
-    static let reuseId: String = "TrackingHistorySectionHeaderView"
-    
-    private var containerView: UIView = {
-        let view = UIView()
-        view.backgroundColor = .white
-        return view
-    }()
-    
-    private var titleLabel: UILabel = {
-        let label = UILabel()
-        label.textAlignment = .left
-        label.font = .amazonFont(type: .bold, size: 13)
-        label.textColor = .black
-        label.numberOfLines = 1
-        label.lineBreakMode = .byWordWrapping
-        return label
-    }()
-    
-    private var separatorView = {
-        let view = UIView()
-        view.backgroundColor = .textFieldBackgroundColor
-        return view
-    }()
-    
-    var title: String? {
-        set { titleLabel.text = newValue }
-        get { titleLabel.text }
-    }
-    
-    override init(reuseIdentifier: String?) {
-        super.init(reuseIdentifier: reuseIdentifier)
-        setupViews()
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        setupViews()
-    }
-    
-    private func setupViews() {
-        self.addSubview(containerView)
-        containerView.addSubview(titleLabel)
-        containerView.addSubview(separatorView)
-        
-        containerView.snp.makeConstraints {
-            $0.top.bottom.leading.trailing.equalToSuperview()
-        }
-        
-        titleLabel.snp.makeConstraints {
-            $0.top.equalToSuperview().offset(20)
-            $0.leading.equalToSuperview().offset(24)
-            $0.trailing.equalToSuperview().offset(-24)
-        }
-        
-        separatorView.snp.makeConstraints {
-            $0.top.equalTo(titleLabel.snp.bottom).offset(20)
-            $0.leading.trailing.bottom.equalToSuperview()
-        }
-    }
-}
