+++ LocationServices/LocationServices/Scenes/Tracking/Subviews/Tracking Simulation/Contracts/TrackingSimulationContracts.swift
@@ -12,13 +12,10 @@ protocol TrackingSimulationViewModelProtocol : AnyObject  {
     func loadData() async
     func startTracking(lat: Double, long: Double)
     func getItemCount(for section: Int) -> Int
-    func getCellModel(indexPath: IndexPath) -> TrackHistoryCellModel?
     func getTitle(for section: Int) -> String
     func sectionsCount() -> Int
-    func setHistory(_ history: [TrackingHistoryPresentation])
     func getTrackingStatus() -> Bool
     func changeTrackingStatus(_ isTrackingActive: Bool)
-    func deleteHistory() async throws
 }
 
 protocol TrackingSimulationViewModelOutputDelegate: AnyObject, AlertPresentable {
